nsim: remove buffer u32 upper bound
Extreme BDP flows requires more buffer memory.

+++ src/plugins/nsim/nsim.c
@@ -188,7 +188,7 @@ nsim_configure (nsim_main_t * nsm, f64 bandwidth, f64 delay, f64 packet_size,
   nsm->drop_fraction = drop_fraction;
 
   /* delay in seconds, bandwidth in bits/sec */
-  total_buffer_size_in_bytes = (u32) ((delay * bandwidth) / 8.0) + 0.5;
+  total_buffer_size_in_bytes = ((delay * bandwidth) / 8.0) + 0.5;
 
   /*
    * Work out how much buffering each worker needs, assuming decent
